requirements.txt
# Core Libraries 
# Environment and Data Handling
python-dotenv  
ipykernel  
streamlit  
requests
numpy
pandas
openai